Refactor process routines out of saml.ts
